+++ rust/dbn/src/encode/csv.rs
@@ -479,7 +479,7 @@ mod tests {
     use super::*;
     use crate::{
         encode::test_data::{VecStream, BID_ASK, RECORD_HEADER},
-        enums::{InstrumentClass, SecurityUpdateAction},
+        enums::{InstrumentClass, SecurityUpdateAction, UserDefinedInstrument},
         record::{
             str_to_c_chars, ImbalanceMsg, InstrumentDefMsg, MboMsg, Mbp10Msg, Mbp1Msg, OhlcvMsg,
             StatusMsg, TradeMsg, WithTsOut,
@@ -596,7 +596,6 @@ mod tests {
             ts_recv: 1658441891000000000,
             ts_in_delta: 22_000,
             sequence: 1_002_375,
-            booklevel: [],
         }];
         let mut buffer = Vec::new();
         let writer = BufWriter::new(&mut buffer);
@@ -719,7 +718,7 @@ mod tests {
             maturity_month: 8,
             maturity_day: 9,
             maturity_week: 11,
-            user_defined_instrument: 'N' as c_char,
+            user_defined_instrument: UserDefinedInstrument::No,
             contract_multiplier_unit: 0,
             flow_schedule_type: 5,
             tick_rule: 0,
@@ -748,7 +747,6 @@ mod tests {
                 ts_recv: 1658441891000000000,
                 ts_in_delta: 22_000,
                 sequence: 1_002_375,
-                booklevel: [],
             },
             ts_out: 1678480044000000000,
         }];