+++ rust/dbn/src/lib.rs
@@ -1,9 +1,10 @@
 //! A crate for reading DBN and legacy DBZ files and converting them to other
 //! [`Encoding`](enums::Encoding)s.
 
-#[deny(missing_docs)]
-#[deny(rustdoc::broken_intra_doc_links)]
-#[deny(clippy::missing_errors_doc)]
+#![cfg_attr(docsrs, feature(doc_auto_cfg))]
+#![deny(missing_docs)]
+#![deny(rustdoc::broken_intra_doc_links)]
+#![deny(clippy::missing_errors_doc)]
 pub mod decode;
 pub mod encode;
 pub mod enums;
@@ -27,3 +28,8 @@ const SYMBOL_CSTR_LEN: usize = 22;
 const NULL_END: u64 = u64::MAX;
 const NULL_LIMIT: u64 = 0;
 const NULL_RECORD_COUNT: u64 = u64::MAX;
+
+/// The sentinel value for an unset or null price.
+pub const UNDEF_PRICE: i64 = i64::MAX;
+/// The sentinel value for an unset or null order quantity.
+pub const UNDEF_ORDER_SIZE: u32 = u32::MAX;