+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.4.1 - 2023-04-05
+- Added enums `MatchAlgorithm`, `UserDefinedInstrument`
+- Added constants `UNDEF_PRICE` and `UNDEF_ORDER_SIZE`
+- Added Python type stubs for `databento_dbn` package
+- Fixed `Metadata.__bytes__` method to return valid DBN
+- Fixed panics when decoding invalid records
+- Fixed issue with attempting to decode partial records in Python `DbnDecoder`
+- Fixed support for `ImbalanceMsg` in Python `DbnDecoder`
+
 ## 0.4.0 - 2023-03-24
 - Added support for Imbalance schema
 - Updated `InstrumentDefMsg` to include options-related fields and `instrument_class`